Implementando validaciones para empleado

+++ src/main/java/com/example/TvBackend/Service/ProductoService.java
@@ -1,11 +1,14 @@
 package com.example.TvBackend.Service;
 
+import com.example.TvBackend.Model.Categoria;
 import com.example.TvBackend.Model.Producto;
 import com.example.TvBackend.Repository.ProductoRepository;
+import com.example.TvBackend.Utilidades.Utilidades;
 import com.example.TvBackend.constantes.Constantes;
 import com.example.TvBackend.interfaceService.IProductoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -15,6 +18,11 @@ public class ProductoService implements IProductoService {
 
     @Autowired
     private ProductoRepository productoDao;
+
+    @Override
+    public List<Producto> buscarPorNombreOMarca(String filtro) {
+        return productoDao.getProductosByNombreOrMarcaOrderByNombreDesc(filtro);
+    }
 
     @Override
     public List<Producto> obtenerProductos() {
@@ -39,6 +47,7 @@ public class ProductoService implements IProductoService {
             productoActualizado.setCaracteristicas(product.getCaracteristicas());
             productoActualizado.setGarantia(product.getGarantia().toUpperCase());
             productoActualizado.setReferencia(product.getReferencia().toUpperCase());
+            productoActualizado.setCategoria(product.getCategoria());
             productoDao.save(productoActualizado);
             return producto;
         } else {
@@ -47,41 +56,35 @@ public class ProductoService implements IProductoService {
     }
 
     @Override
-    public Producto registrarProducto(Map<String, String> producto) {
+    public ResponseEntity<String> registrarProducto(Producto productoJson) {
         try {
-            if (validarInformacion(producto)) {
-                return productoDao.save(getFromMapProducto(producto));
-            } else {
-                throw new NoSuchElementException("Datos invalidos" + HttpStatus.BAD_REQUEST);
-            }
+            productoDao.save(convertirAtributosToMayuscula(productoJson));
+            return Utilidades.getResponseEntity("Producto guardado con exito. Status:" + HttpStatus.OK, HttpStatus.OK);
         } catch (Exception e) {
             e.printStackTrace();
         }
-        throw new NoSuchElementException(Constantes.ALGO_PASO_MAL + HttpStatus.BAD_REQUEST);
+        return Utilidades.getResponseEntity("Ocurrio un error interno en el servidor", HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @Override
-    public void eliminarProducto(int id) {
-        productoDao.deleteById(id);
-    }
-
-    public boolean validarInformacion(Map<String, String> producto) {
-        if (producto.containsKey("nombre") && producto.containsKey("marca") && producto.containsKey("costo")
-                && producto.containsKey("caracteristicas") && producto.containsKey("referencia")) {
-            return true;
+    public ResponseEntity<String> eliminarProducto(int id) {
+        Long id_long = (long) id;
+        if (id_long != null) {
+            productoDao.deleteById(id);
+            return Utilidades.getResponseEntity("La categoria con el id: " + id + " se eliminó con exito", HttpStatus.OK);
         }
-        return false;
+        return Utilidades.getResponseEntity("No se encontró la categoria con el ID especificado", HttpStatus.BAD_REQUEST);
     }
-
-    public Producto getFromMapProducto(Map<String, String> productoMap) {
-        Producto producto = new Producto();
-        producto.setNombre(productoMap.get("nombre").toUpperCase());
-        producto.setFoto((productoMap.get("foto")));
-        producto.setMarca(productoMap.get("marca").toUpperCase());
-        producto.setCosto(Float.parseFloat(productoMap.get("costo")));
-        producto.setCaracteristicas(productoMap.get("caracteristicas"));
-        producto.setGarantia(productoMap.get("garantia").toUpperCase());
-        producto.setReferencia(productoMap.get("referencia").toUpperCase());
-        return producto;
+    public Producto convertirAtributosToMayuscula(Producto producto){
+        Producto productoEnMayus = new Producto();
+        productoEnMayus.setMarca(producto.getNombre().toUpperCase());
+        productoEnMayus.setMarca(producto.getMarca().toUpperCase());
+        productoEnMayus.setFoto(producto.getFoto());
+        productoEnMayus.setCosto(producto.getCosto());
+        productoEnMayus.setCaracteristicas(producto.getCaracteristicas().toUpperCase());
+        productoEnMayus.setGarantia(producto.getGarantia());
+        productoEnMayus.setReferencia(producto.getReferencia().toUpperCase());
+        productoEnMayus.setCategoria(producto.getCategoria());
+        return productoEnMayus;
     }
 }
